+++ lib/js_of_ocaml/abort.mli
@@ -1,6 +1,5 @@
-(* Js_of_ocaml compiler
+(* Js_of_ocaml library
  * http://www.ocsigen.org/js_of_ocaml/
- * Copyright (C) 2020 Hugo Heuzard
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,48 +16,32 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
 
-let runtime =
-  Files.
-    [ array
-    ; backtrace
-    ; bigarray
-    ; bigstring
-    ; compare
-    ; fail
-    ; format
-    ; fs
-    ; fs_fake
-    ; fs_node
-    ; gc
-    ; graphics
-    ; hash
-    ; ieee_754
-    ; float32
-    ; int64
-    ; ints
-    ; io
-    ; jslib
-    ; jslib_js_of_ocaml
-    ; lexing
-    ; marshal
-    ; md5
-    ; mlBytes
-    ; nat
-    ; obj
-    ; parsing
-    ; promise
-    ; stdlib
-    ; sys
-    ; str
-    ; unix
-    ; weak
-    ; domain
-    ; prng
-    ; sync
-    ; effect_
-    ; zstd
-    ; runtime_events
-    ; blake2
-    ]
+(** AbortController / AbortSignal.
 
-include Files
+    A general-purpose cancellation primitive used by {!Fetch}, event
+    listener registration, [Streams], and custom asynchronous code.
+
+    @see <https://developer.mozilla.org/en-US/docs/Web/API/AbortController>
+    @see <https://dom.spec.whatwg.org/#interface-abortcontroller> *)
+
+open Js
+
+class type signal = object ('self)
+  method aborted : bool t readonly_prop
+
+  method reason : Unsafe.any readonly_prop
+
+  method onabort : ('self t, 'self Dom.event t) Dom.event_listener writeonly_prop
+
+  method throwIfAborted : unit meth
+end
+
+class type controller = object
+  method signal : signal t readonly_prop
+
+  method abort : unit meth
+
+  method abort_reason : Unsafe.any -> unit meth
+end
+
+val controller : controller t constr